Update patch set 1
Patch Set 1:

(1 comment)

I don't think devices use it now, but it stays as a reference that is / was locally forked in devices.

I was planning to add the related "add support to capture ip table stats" update to sonyxperiadev, but had to fix the issues it introduces first.

Patch-set: 1
Label: Verified=0