Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The fix is simple: remove rmnetcli from your target packages. Merging this commit can cause potential merge conflicts when merging new tags from CAF, for absolutely no real reason. There is no need to further discuss this.
If it really is useless and unused as you think, it will never be changed, so it can not lead to merge conflicts.
Only case I can think of, would be if it gets removed at all.
But I see, that there is no need for further discussion.
I really appreciate what lineageos did and does for the android community, but sometimes, I can not understand the attitude of it's developers.
I'm out and wish everybody a good night or a good day, depending on your location.

Patch-set: 2
Reviewer: Gerrit User 21506 <21506@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>